add all

+++ src/main/java/net/davidvoid/thor/lightning/service/GroupService.java
@@ -2,10 +2,12 @@ package net.davidvoid.thor.lightning.service;
 
 import java.util.List;
 
+import net.davidvoid.thor.lightning.data.access.FeedStore;
 import net.davidvoid.thor.lightning.data.access.GroupStore;
 import net.davidvoid.thor.lightning.data.access.UserStore;
 import net.davidvoid.thor.lightning.entity.Group;
 import net.davidvoid.thor.lightning.entity.User;
+import net.davidvoid.thor.lightning.exception.ResourceBusyException;
 import net.davidvoid.thor.lightning.util.SessionUser;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +20,8 @@ public class GroupService {
     GroupStore store = null;
     @Autowired
     UserStore userStore = null;
+    @Autowired
+    FeedStore feedStore = null;
     
     public List<Group> getAllGroups() {
         return store.getGroups(userInSession());
@@ -36,6 +40,10 @@ public class GroupService {
     }
 
     public void delete(Object group_id) {
+        if (feedStore.countFeedsByGroupId(group_id) != 0) {
+            throw new ResourceBusyException("group can't be deleted because it contains feeds");
+        }
+
         store.deleteById(group_id);
     }
